server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
youtube-transcript-api>=0.6.1
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pdf2image>=1.16.0
Pillow>=10.0.0
